+++ src/Resources/Doc/changelog.md
@@ -17,6 +17,11 @@ Version 1.0
 ## Version 0.15
 * Foundational work on Dose functionality.
 
+#### v0.15.21200.1627 (2021-07-19)
+* `INFO` Clean build for testing.
+##### Dose
+* `ADDED` formLoopCount and fieldLoopCount for debugging purposes.
+
 #### v0.15.21200.1537 (2021-07-19)
 * `INFO` Clean build for testing.